Move database files
