+++ src/components/actions/getApplicationData.ts
@@ -1,5 +1,5 @@
 import { ActionApplicationData, ActionPayload, BaseApplicationData, ReviewData } from '../../types'
-import DBConnect from '../databaseConnect'
+import DBConnect from '../database/databaseConnect'
 import { BasicObject } from '../../modules/expression-evaluator'
 import { getAppEntryPointDir } from '../utilityFunctions'
 import config from '../../config'
